requirements.txt
# XHS-OCR 项目依赖
# 基础依赖
opencv-python>=4.12.0.88
numpy>=2.2.6
Pillow>=12.0.0
pymysql>=1.1.2
python-dotenv>=1.1.1
# 定时任务（可选）
schedule>=1.2.2
# 配置文件处理
configparser>=7.2.0
loguru>=0.7.3
playwright>=1.55.0
#accelerate==1.11.0
#einops==0.8.1
#fsspec==2025.9.0
torch==2.9.0
torchvision
surya-ocr==0.17.0

#safetensors==0.6.2
#sentencepiece==0.2.1
#sniffio==1.3.1
#starlette==0.48.0
#tokenizers==0.22.1
#torch==2.9.0
#torchvision==0.24.0
#transformers==4.57.1
